Skipping download if already exists

+++ scripts/preprocess.sh
@@ -1,25 +1,37 @@
 #!/bin/bash
 dataset_url="https://disk.yandex.com/d/NdpjIZDmP24kng"
 
-# Check if yadisk-direct is installed
-if command -v yadisk-direct >/dev/null 2>&1; then
-  dataset_direct_url="$(yadisk-direct $dataset_url)"
-else
-  echo "Command yadisk-direct does not exist! Please, run 'pip install -r requirements.txt'"
-  exit
+already_exists=false
+
+# Check if dataset already exists
+if [ -f data/raw/vehicles.csv ]; then
+  already_exists=true
+  echo "Dataset already exists! Skipping downloading..."
 fi
 
-# Pull dataset from the Yandex.Disk
-if wget $dataset_direct_url -O data/usedcars.zip; then
-  echo "Success!"
-else
-  echo "Something went wrong!"
-  exit
-fi
+if [ "$already_exists" = false ]; then
 
-# Unzip file and remove artifacts
-unzip -o data/usedcars.zip -d data/raw/
-rm data/usedcars.zip
+  # Check if yadisk-direct is installed
+  if command -v yadisk-direct >/dev/null 2>&1; then
+    dataset_direct_url="$(yadisk-direct $dataset_url)"
+  else
+    echo "Command yadisk-direct does not exist! Please, run 'pip install -r requirements.txt'"
+    exit
+  fi
+
+  # Pull dataset from the Yandex.Disk
+  if wget $dataset_direct_url -O data/usedcars.zip; then
+    echo "Success!"
+  else
+    echo "Something went wrong!"
+    exit
+  fi
+
+  # Unzip file and remove artifacts
+  unzip -o data/usedcars.zip -d data/raw/
+  rm data/usedcars.zip
+
+fi
 
 echo "Data is ready!"
